UpscalePi/requirements.txt
pinecone==7.0.1
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==4.1.0
tiktoken==0.9.0
